Split hosted connection implementation modules
